реализация работающего клиентся с сервером с корректным отключением клиента
