+++ lesson6/src/Client/Main.java
@@ -1,0 +1,53 @@
+package Client;
+
+import javafx.application.Application;
+import javafx.event.EventHandler;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
+import javafx.scene.image.Image;
+import javafx.scene.input.MouseEvent;
+import javafx.scene.paint.Color;
+import javafx.stage.Stage;
+import javafx.stage.StageStyle;
+
+
+public class Main extends Application {
+    private double xOffset;
+    private double yOffset;
+
+
+        @Override
+        public void start ( final Stage stage) throws Exception {
+            Parent root = FXMLLoader.load(getClass().getResource("sample.fxml"));
+            stage.initStyle(StageStyle.UNDECORATED);
+            stage.setTitle("MyChat");
+            stage.getIcons().add(new Image("file:images/iconfinder.png"));
+                       Scene scene = new Scene(root, 300, 275, Color.TRANSPARENT);
+            scene.setOnMousePressed(new EventHandler<MouseEvent>() {
+                @Override
+                public void handle(MouseEvent event) {
+                    xOffset = stage.getX() - event.getScreenX();
+                    yOffset = stage.getY() - event.getScreenY();
+                }
+            });
+            scene.setOnMouseDragged(new EventHandler<MouseEvent>() {
+                @Override
+                public void handle(MouseEvent event) {
+                    stage.setX(event.getScreenX() + xOffset);
+                    stage.setY(event.getScreenY() + yOffset);
+                }
+            });
+
+            stage.setScene(scene);
+            stage.initStyle(StageStyle.UNDECORATED);
+            stage.setResizable(false);
+            stage.show();
+        }
+
+
+
+    public static void main(String[] args) {
+        launch(args);
+    }
+}
